Arquitectura/Cloud Function/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-secret-manager
nbformat
nbconvert
google-api-python-client